Test 3 passess

+++ app/configuration.c
@@ -1,8 +1,8 @@
 struct configuration {
 
     int N;
-	double Tau;
-	double T;
-	double alpha;
-	double delta;
+	long double Tau;
+	long double T;
+	long double alpha;
+	long double delta;
 };
